# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AfyaCommerce Internal API
  version: 1.0.0
  description: Text-first API contract for patient and rider mobile clients.
servers:
  - url: /
    description: Same-origin internal API
security:
  - bearerAuth: []
paths:
...
  /api/mpesa/stk-push:
    post:
      summary: Initiate an M-PESA STK push
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StkPushRequest"
      responses:
        "200":
          description: Checkout request
          content:
            application/json:
              schema:
                type: object
                required: [CheckoutRequestID]
                properties:
                  CheckoutRequestID:
                    type: string
                  CustomerMessage:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    StkPushRequest:
      type: object
      required: [order_id, order_number, phone, amount]
      properties:
        order_id:
          type: string
        order_number:
          type: string
        phone:
          type: string
        amount:
          type: number
          format: double
          minimum: 1
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
...